End of forward_loop beginning of forward_char_loop
In this code segment we ended the initial forward loop segment and then began working on the character loop. The forward loop was used to ensure we correctly calculate the string index and also keep track of what character we are on now were are working on calculating the position the character is in the circular loop. As each character is technically going in a loop.